refactor: remove magic numbers and unnecessary enum casts
- effect_reg_footprint.mm: EffectTypeFootprint instead of magic number 1
- effect_reg_pomodorobed.mm: EffectTypePomodoroBed instead of magic number 5
- effect_window.mm: removed 4 unnecessary (int) casts from enum comparisons
- registration struct stays int (EffectType lives in effect_window.h, circular dep)
- initWithType: cast kept (Objective-C method signature requires it)

+++ src/platform/macos/effect_reg_pomodorobed.mm
@@ -57,7 +57,7 @@ static void PomodoroBed_ConfigureWindow(EffectWindow* win, const Vector2& pos) {
 static struct PomodoroBedRegistration {
     PomodoroBedRegistration() {
         EffectRegister({
-            (int)EffectTypePomodoroBed,
+            EffectTypePomodoroBed,
             PomodoroBed_GetPositions,
             PomodoroBed_GetRadius,
             PomodoroBed_ExistsAt,
